+++ src/types/address.types.ts
@@ -1,0 +1,4 @@
+export type LatLng = {
+  La: number;
+  Ma: number;
+};
