+++ tsconfig.json
@@ -26,7 +26,7 @@
     "**/*.ts",
     "**/*.tsx",
     ".next/types/**/*.ts",
-    "src/app/(provider)/(root)/sponsored-store/page.tsx"
+    "src/app/(provider)/(root)/free-meals/map/page.tsx"
   ],
   "exclude": ["node_modules"]
 }
